Introduced more advanced profile viewing and editing. Separated form into it's own class.

+++ code/cms/MemberBulkLoader.php
@@ -15,16 +15,20 @@ class MemberBulkLoader extends MemberCsvBulkLoader {
 		$obj->Surname = join(' ', $nameParts);
 	}
 	
-	protected function processRecord($record, $columnMap, &$results, $preview = false) {
-			$this->extend('preprocess', $record, $columnMap, $results, $preview);
-			$id = parent::processRecord($record, $columnMap, $results, $preview);
-			if($member = Member::get()->byID($id)){
-				$member->write();
-				$this->extend('postprocess', $member, $record, $columnMap, $results, $preview);//callback for doing other custom stuff
-				$member->destroy();
-				unset($member);
-			}
-			return $id;
+	public function processRecord($record, $columnMap, &$results, $preview = false) {
+		$skip = false;
+		$this->extend('preprocess', $record, $columnMap, $results, $preview, $skip);
+		if($skip){
+			return 0;
+		}
+		$id = parent::processRecord($record, $columnMap, $results, $preview);
+		if($member = Member::get()->byID($id)){
+			$this->extend('postprocess', $member, $record, $columnMap, $results, $preview);//callback for doing other custom stuff
+			$member->write();
+			$member->destroy();
+			unset($member);
+		}
+		return $id;
 	}
 
 }
